requirements.txt
annotated-types==0.7.0
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi @ file:///croot/certifi_1700501669400/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cmake==3.30.2
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1702070282333/work
cycler==0.12.1
dgl==1.1.3
einops==0.7.0
embed==0.1.0
et-xmlfile==1.1.0
filelock==3.13.4
finch-clust==0.2.0
fonttools==4.47.0
fsspec==2024.3.1
huggingface-hub==0.22.2
idna @ file:///croot/idna_1666125576474/work
importlib-resources==6.1.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lit==18.1.8
MarkupSafe==2.1.5
matplotlib==3.7.4
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mlxtend==0.23.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
openpyxl==3.1.5
packaging==24.0
pandas==2.0.3
PATE==0.1.1
patsy==0.5.6
Pillow @ file:///croot/pillow_1696580024257/work
psutil==5.9.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.7.3
pydantic_core==2.18.4
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests @ file:///croot/requests_1690400202158/work
safetensors==0.4.3
SciencePlots==1.0.9
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
statsmodels==0.14.1
sympy==1.12.1
threadpoolctl==3.5.0
tokenizers==0.15.2
torch==2.0.0
torchaudio==0.11.0
torchdata==0.7.1
torchvision==0.12.0
tqdm==4.66.2
transformers==4.39.3
triton==2.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3 @ file:///croot/urllib3_1698257533958/work
xlrd==1.2.0
zipp==3.17.0
